# Optimize parsing of parameters in include tag