## Summary

- Replace `while..do` loop generating `MatchData` with *faster* `String#scan` and a block to use meaningful local variable names.
- Since `String#gsub` duplicates the input string irrespective of a match, test if substitution is necessary before calling `String#gsub`. *(The overhead for prior testing is negligible.)*