@DirtyF 2020-05-20
How can we measure the gain here?