@ashmaroli 2020-05-20
> How can we measure the gain here?

@DirtyF I've added a benchmark script to this branch as well.. You can run it by checking the PR branch out locally:

```sh
git fetch upstream pull/8192/head:optimize-includes-params-parsing
git checkout optimize-includes-params-parsing
ruby benchmark/parse-include-tag-params.rb
```